requirements.txt
autopep8==1.5.7
flake8==3.9.2
flask==2.0.1
gunicorn==20.1.0
flask-wtf==0.15.1
flask-sqlalchemy==2.5.1
psycopg2-binary==2.8.6
argon2-cffi
